[BEAM-2613] add integration test for comparison operators
